fixed: CSSHelperWebkit not working for Chrome 40+
"style.getPropertyCSSValue" is undefined for the current version of Chrome. It seems they removed support for that call. My attempt to fix this broken behaviour was to simply call:

return style.getPropertyValue(name);

This seems to fix the problems we currently have in our application. I cannot accurately assess whether the problem is correctly and completely solved with this fix. According to the former written comment, there might be a problem with "named colors", but I do not know how to prove that!