Fix to bug when using SVGImage with ui:binder
In the original version the following ui:binder code doesn’t work

<svg:SVGImage ui:field="fbutton" resource="{commonRes.iconF}" width="32px"/>

since it seems that gwt, on ui creation, tries to set width before setting the resource, which triggers a null pointer exception. By declaring the constructor using @UIConstructor tag, gwt sets the specified resource on construction, which is before setting the other attributes specified in the ui:binder tag (in the example, the attribute ‘width’).